chore: Used query wrappers more improved resuability
