+++ app/hotels/components/filters/LocationPicker.jsx
@@ -1,8 +1,7 @@
 "use client";
 
 import LocationPicker from "@/app/components/search-filters/LocationPicker";
-import { useQuery } from "@tanstack/react-query";
-import { getLocationsByName } from "../../queryFunctions";
+import { useGetLocationByName } from "@/hooks/query-hooks/useGetLocationByName";
 import { observer, useObservable } from "@legendapp/state/react";
 import debounce from "debounce";
 
@@ -10,13 +9,7 @@ const HotelLocationPicker = observer(function Component() {
   const textSearchTerm$ = useObservable("");
   const textSearchTerm = textSearchTerm$.get();
 
-  const { data: locations, refetch } = useQuery({
-    queryKey: ["locations", textSearchTerm],
-    queryFn: () => getLocationsByName(textSearchTerm),
-    enabled: !!textSearchTerm,
-    select: (data) => data?.hotel_listing_locations,
-    staleTime: 1000 * 60 * 15, // store for up to 15 minutes
-  });
+  const [locations, refetch] = useGetLocationByName (textSearchTerm);
 
   const handleLocationSearch = debounce((searchTerm) => {
     textSearchTerm$.set(searchTerm);
